# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProyectoTech",
        "description": "Documentación proyectoTech (CRUD de productos y carga masiva - Carrito de Compras)",
        "contact": {
            "email": "johan33@hotmail.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Desarrollo ProyectoTech"
        }
    ],
    "paths": {
...
        "/auth/reporte/ventas": {
            "post": {
                "tags": [
                    "Informes"
                ],
                "summary": "Ventas por fecha",
                "description": "Servicio para consultar las ventas realizadas de acuerdo al rango de fechas. <b>El informe lo encontrarán en la carpeta storage/app/public del proyecto</b>",
                "operationId": "reporte/ventas",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "inicio": {
                                        "description": "Fecha inicial de ventas (YYYY-MM-DD)",
                                        "type": "string",
                                        "example": "2022-02-17"
                                    },
                                    "fin": {
                                        "description": "Fecha final de ventas (YYYY-MM-DD)",
                                        "type": "string",
                                        "example": "2022-02-20"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transacción exitosa",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "400": {
                        "description": "Petición no válida"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    },
                    "500": {
                        "description": "Error de servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}